# resources/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BWH eSign native API",
    "version": "1.0.0",
    "summary": "Prepare, send, track, and retrieve signed PDF agreements.",
    "description": "The native HTTP surface of BWH eSign. It and the Firma-compatible facade under `/functions/v1/signing-request-api` call the same domain services; signing rules live in the state machine and never in an HTTP controller.\n\n**Tenancy.** There is no workspace parameter anywhere in this API. The credential presented in `Authorization` belongs to exactly one workspace, and that is the tenant for every lookup. An id belonging to another workspace answers `404`, identical to an id that does not exist, so a caller cannot probe for the existence of another tenant's records.\n\n**Errors.** Every failure is `{\"error\": {\"code\", \"message\", \"details\"}}`. Switch on `code`; the message is for a human reading a log. Codes are stable and renaming one is a breaking change.\n\n**Honest language.** Humans provide electronic signatures and assent; the service seals the finished PDF under its own certificate. Nothing here is a per-signer certificate, and no assurance level beyond the PAdES baselines named in `assurance_level` is claimed.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This deployment."
    }
  ],
  "security": [
    {
      "serviceCredential": []
    }
  ],
...
  "paths": {
...
    "/envelopes/{envelope}/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "listEnvelopeEvents",
        "summary": "The envelope's events, oldest first.",
        "description": "The pull half of the webhook feed, with a cursor a caller can persist. An integration that cannot accept inbound HTTP, or whose receiver was down, reads the same events here in the same order. `id` is the event id a webhook receiver deduplicates on.",
        "parameters": [
          {
            "name": "envelope",
            "in": "path",
            "required": true,
            "description": "The envelope's public id (ULID).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor from a previous response's `meta.next_cursor`. Omit for the first page. A value this API did not issue is `422 invalid_cursor`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size. Clamped to 100 rather than rejected, because a client asking for more wants as many as it can have.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of events.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No envelope with that id exists in this credential's workspace. Another tenant's id is indistinguishable from one that does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The cursor is not one this API issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one that is unknown, revoked, or expired. `error.code` is `invalid_credential`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The credential is valid but was not granted the scope this endpoint requires. `error.code` is `insufficient_scope`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected failure. The body carries no stack trace, no file path, and no SQL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceCredential": {
        "type": "http",
        "scheme": "bearer",
        "description": "A service credential in `Authorization`. Both syntaxes authenticate and are equivalent:\n\n    Authorization: esk_<prefix>_<secret>\n    Authorization: Bearer esk_<prefix>_<secret>\n\nThe raw form is accepted because the API this application is compatible with sends its key with no scheme. Native callers should use `Bearer`. Credentials are issued, rotated, and revoked from the console (`docs/operations/service-credentials.md`); there is no self-service route and no way to read a secret back."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "description": "Every failure on this API. Switch on `error.code`; the message is for a human reading a log.",
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "invalid_credential",
                  "insufficient_scope",
                  "not_found",
                  "method_not_allowed",
                  "illegal_transition",
                  "conflict",
                  "not_completed",
                  "idempotency_key_in_flight",
                  "validation_failed",
                  "idempotency_key_reused",
                  "invalid_cursor",
                  "send_preconditions_failed",
                  "field_rejected",
                  "invalid_snapshot",
                  "unknown_recipient",
                  "template_state",
                  "destination_refused",
                  "unknown_event_name",
                  "unsupported",
                  "internal_error"
                ],
                "description": "Stable machine code. Renaming one is a breaking change."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "Present when there is something structured to say: `fields` for validation, `problems` for a send gate, `required_scope` for a 403."
              }
            }
          }
        }
      },
      "PageMeta": {
        "type": "object",
        "required": [
          "next_cursor"
        ],
        "description": "There is no total and no page number. Counting a tenant's rows on every request buys a number that is stale before it renders, and there are no pages to number in a keyset scheme.",
        "properties": {
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass back as `?cursor=` for the next page. Null means this is the last."
          }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The event identity a webhook receiver deduplicates on. The same id appears in the delivered body."
          },
          "event": {
            "type": "string",
            "examples": [
              "signing_request.sent"
            ],
            "description": "Names from the compatibility profile are reproduced verbatim; anything of ours is prefixed `esign.`."
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the transition happened, not when this row was read."
          },
          "payload": {
            "type": "object",
            "additionalProperties": true,
            "description": "The event's data object, exactly as recorded. Never rebuilt from current state."
          }
        }
      },
...
    }
  }
}